requirements.txt
dbt-core==1.8.*
dbt-duckdb==1.8.*
dbt-snowflake==1.8.*
duckdb>=0.10.0
pandas>=2.0
sqlfluff==3.0.*
apache-airflow==2.9.*
